+++ tests/integration/test_CreativeProject_ask_tell_functional.py
@@ -1,7 +1,7 @@
 import functools
 import pytest
 import torch
-from greattunes import CreativeProject
+from greattunes import TuneSession
 from greattunes.data_format_mappings import tensor2pretty_response, tensor2pretty_covariate
 
 
@@ -20,11 +20,11 @@ from greattunes.data_format_mappings import tensor2pretty_response, tensor2prett
 def test_CreativeProject_ask_integration_test_works(covars, model_type, train_X, train_Y, covars_proposed_iter,
                                                     covars_sampled_iter, response_sampled_iter, random_start):
     """
-    test the positive cases for CreativeProject.ask method.
+    test the positive cases for TuneSession.ask method.
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type, random_start=random_start)
+    cc = TuneSession(covars=covars, model=model_type, random_start=random_start)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality)
     cc.train_X = train_X
@@ -69,12 +69,12 @@ def test_CreativeProject_ask_integration_test_works(covars, model_type, train_X,
 def test_CreativeProject_ask_integration_test_fails(covars, model_type, train_X, train_Y, covars_proposed_iter,
                                                     covars_sampled_iter, response_sampled_iter, error_msg):
     """
-    test the negative cases for CreativeProject.ask method. Currently testing the case where only train_Y data has been
+    test the negative cases for TuneSession.ask method. Currently testing the case where only train_Y data has been
     added, not train_X data
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type)
+    cc = TuneSession(covars=covars, model=model_type)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality)
     cc.train_X = train_X
@@ -103,12 +103,12 @@ def test_CreativeProject_ask_integration_test_fails(covars, model_type, train_X,
 def test_CreativeProject_tell_integration_test_works(covars, model_type, train_X, train_Y, covars_proposed_iter,
                                                      covars_sampled_iter, response_sampled_iter, monkeypatch):
     """
-    test the positive cases for CreativeProject.tell method. Monkeypatch "_read_covars_manual_input" and
+    test the positive cases for TuneSession.tell method. Monkeypatch "_read_covars_manual_input" and
     "_read_response_manual_input" from ._observe.py to circumvent manual input via builtins.input
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type)
+    cc = TuneSession(covars=covars, model=model_type)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality)
     cc.train_X = train_X
@@ -169,13 +169,13 @@ def test_CreativeProject_tell_integration_test_works(covars, model_type, train_X
 def test_CreativeProject_tell_integration_test_works_overwrite(covars, model_type, train_X, train_Y, covars_proposed_iter,
                                                      covars_sampled_iter, response_sampled_iter, monkeypatch):
     """
-    test the positive case for CreativeProject.tell method where last datapoint is overwritten (controlled by
+    test the positive case for TuneSession.tell method where last datapoint is overwritten (controlled by
     covars_proposed_iter <= covars_sampled_iter). Monkeypatch "_read_covars_manual_input" and
     "_read_response_manual_input" from ._observe.py to circumvent manual input via builtins.input
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type)
+    cc = TuneSession(covars=covars, model=model_type)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality)
     cc.train_X = train_X
@@ -245,7 +245,7 @@ def test_CreativeProject_tell_integration_test_fails(covars, model_type, train_X
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type)
+    cc = TuneSession(covars=covars, model=model_type)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality)
     cc.train_X = train_X
@@ -309,7 +309,7 @@ def test_CreativeProject_integration_ask_tell_one_loop_works(covars, model_type,
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type, random_start=random_start)
+    cc = TuneSession(covars=covars, model=model_type, random_start=random_start)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality)
     cc.train_X = train_X
@@ -417,7 +417,7 @@ def test_CreativeProject_integration_ask_tell_one_loop_kwarg_covars_works(covars
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type, random_start=random_start)
+    cc = TuneSession(covars=covars, model=model_type, random_start=random_start)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality)
     cc.train_X = train_X
@@ -506,7 +506,7 @@ def test_CreativeProject_integration_ask_tell_one_loop_kwarg_response_works(cova
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type, random_start=random_start)
+    cc = TuneSession(covars=covars, model=model_type, random_start=random_start)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality)
     cc.train_X = train_X
@@ -592,7 +592,7 @@ def test_CreativeProject_integration_ask_tell_one_loop_kwarg_covars_response_wor
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type, random_start=random_start)
+    cc = TuneSession(covars=covars, model=model_type, random_start=random_start)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality)
     cc.train_X = train_X
@@ -667,7 +667,7 @@ def test_CreativeProject_integration_ask_tell_one_loop_kwarg_covars_fails(covars
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type, random_start=False)
+    cc = TuneSession(covars=covars, model=model_type, random_start=False)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality)
     cc.train_X = train_X
@@ -721,7 +721,7 @@ def test_CreativeProject_integration_ask_tell_one_loop_kwarg_response_fails(cova
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type, random_start=False)
+    cc = TuneSession(covars=covars, model=model_type, random_start=False)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality)
     cc.train_X = train_X
@@ -766,7 +766,7 @@ def test_CreativeProject_integration_ask_tell_ask_works(covars, model_type, trai
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type, random_start=False)
+    cc = TuneSession(covars=covars, model=model_type, random_start=False)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality)
     cc.train_X = train_X
@@ -864,7 +864,7 @@ def test_CreativeProject_integration_ask_ask_tell_overwrite_candidate_works(cova
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type, random_start=False)
+    cc = TuneSession(covars=covars, model=model_type, random_start=False)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality). That is, set attributes set
     # both by _Initializers__initialize_training_data and by _Initializers__initialize_random_start
@@ -951,7 +951,7 @@ def test_CreativeProject_integration_ask_tell_tell_overwrite_covar_resp_works(co
     """
 
     # initialize the class
-    cc = CreativeProject(covars=covars, model=model_type, random_start=False)
+    cc = TuneSession(covars=covars, model=model_type, random_start=False)
 
     # ISSUE IS THAT I AM CIRCUMVENTING THE INITIALIZATION OF RANDOM POINTS WHICH REQUIRES THAT TRAIN_X, TRAIN_Y
     # INITIALIZATION HAS FINISHED. I NEED TO SET RANDOM INITIALIZATION PARAMTERES MANUALLY BELOW
@@ -1063,7 +1063,7 @@ def test_CreativeProject_integration_ask_tell_ask_works(covars, model_type, trai
 
     # initialize the class
     # random_start = True is default, so this tests random start
-    cc = CreativeProject(covars=covars, model=model_type)
+    cc = TuneSession(covars=covars, model=model_type)
 
     # set attributes on class (to simulate previous iterations of ask/tell functionality)
     cc.train_X = train_X
@@ -1180,8 +1180,8 @@ def test_CreativeProject_integration_ask_tell_ask_tell_randon_start_works(train_
     num_initial_random = 1
 
     # initialize the class
-    cc = CreativeProject(covars=covars, train_X=train_X, train_Y=train_Y, random_start=True,
-                         random_sampling_method=random_sampling_method, num_initial_random=num_initial_random)
+    cc = TuneSession(covars=covars, train_X=train_X, train_Y=train_Y, random_start=True,
+                     random_sampling_method=random_sampling_method, num_initial_random=num_initial_random)
 
     # define decorator to add 1.0 to all entries in monkeypatched returned data. This to be able to tell that the last
     # entry (from second "tell") is different than the first, and know that it has been overwritten